[WARNING] Isn't finished, may not work properly. Refactor try
Just to get the idea of what I was trying to do
