+++ include/objects/flow.h
@@ -15,6 +15,8 @@ enum CardinalType {
 };
 typedef enum CardinalType CardinalType;
 
+DxDy cardinal_to_dxdy(CardinalType card);
+
 enum FlowType {
 	ELECTRIC_CABLE,
 	FLUID,
@@ -48,7 +50,7 @@ struct Cable {
 	TileCoords pos;
 	CardinalType connections[2];
 	int maxFlow;
-	int capacity;	
+	int capacity;
 	bool powered;
 };
 typedef struct Cable Cable;
