+++ src/main.c
@@ -6,7 +6,7 @@
 #include <string.h>
 
 #include <SDL2/SDL.h>
-#include <SDL2/SDL_ttf.h>
+// #include <SDL2/SDL_ttf.h>
 
 #include "renderer.h"
 #include "map.h"
@@ -14,242 +14,17 @@
 #include "widget.h"
 #include "entity.h"
 #include "ui.h"
-
-TileCoords window_pixel_to_tile_coords(WinCoords wc) {
-	float tile_render_size = TILE_SIZE * g_camera.zoom;
-	return (TileCoords){
-		.x = floorf(((float)wc.x + g_camera.pos.x) / tile_render_size),
-		.y = floorf(((float)wc.y + g_camera.pos.y) / tile_render_size),
-	};
-}
-
-WinCoords tile_coords_to_window_pixel(TileCoords tc) {
-	float tile_render_size = TILE_SIZE * g_camera.zoom;
-	return (WinCoords){
-		.x = tc.x * tile_render_size - g_camera.pos.x,
-		.y = tc.y * tile_render_size - g_camera.pos.y,
-	};
-}
-
-void render_tile_ground(TileType tile_type, SDL_Rect dst_rect) {
-	SDL_Rect rect_in_spritesheet = g_tile_type_spec_table[tile_type].rect_in_spritesheet;
-	SDL_RenderCopy(g_renderer, g_spritesheet, &rect_in_spritesheet, &dst_rect);
-}
-
-void render_human(SDL_Rect dst_rect) {
-	SDL_Rect rect_in_spritesheet = {0, 16, 3, 4};
-	SDL_RenderCopy(g_renderer, g_spritesheet, &rect_in_spritesheet, &dst_rect);
-}
-
-void render_tile_building(Building* building, SDL_Rect dst_rect) {
-	if (building != NULL){
-		SDL_Rect rect_in_spritesheet;
-		switch (building->type)
-		{
-		case BUILDING_EMITTER:
-			rect_in_spritesheet = g_building_type_spec_table[BUILDING_TX_EMITTER].rect_in_spritesheet;
-			break;
-		case BUILDING_RECEIVER:
-			if (building->powered){
-				rect_in_spritesheet = g_building_type_spec_table[BUILDING_TX_RECEIVER_ON].rect_in_spritesheet;
-			} else {
-				rect_in_spritesheet = g_building_type_spec_table[BUILDING_TX_RECEIVER_OFF].rect_in_spritesheet;
-			}
-			break;
-		default:
-			break;
-		}
-		SDL_RenderCopy(g_renderer, g_spritesheet_buildings, &rect_in_spritesheet, &dst_rect);
-	} else {
-		printf("[ERROR] Missing building !");
-	}
-}
-
-void render_tile_cable(Cable* cable, SDL_Rect dst_rect) {
-	if (cable != NULL){
-		SDL_Rect rect_in_spritesheet;
-		int angle = 0;
-		bool straight = cable->connections[1] - cable->connections[0] == 2;
-			// Cringe af but is just to test. The true way of storing and evaluating directions must be discussed.
-			if (straight){
-				rect_in_spritesheet = g_cable_type_spec_table[ELECTRICITY_STRAIGHT].rect_in_spritesheet;
-				if (cable->powered){
-					rect_in_spritesheet = g_cable_type_spec_table[ELECTRICITY_STRAIGHT_ON].rect_in_spritesheet;
-				}
-
-				angle = 90 * (cable->connections[0] == SOUTH);
-			} else {
-				rect_in_spritesheet = g_cable_type_spec_table[ELECTRICITY_TURN].rect_in_spritesheet;
-				if (cable->powered){
-					rect_in_spritesheet = g_cable_type_spec_table[ELECTRICITY_TURN_ON].rect_in_spritesheet;
-				}
-				angle = 90 * (4-cable->connections[1]) * !(cable->connections[0] == WEST && cable->connections[1] == NORTH);
-			}
-			SDL_RenderCopyEx(g_renderer, g_spritesheet_buildings, &rect_in_spritesheet, &dst_rect, angle, NULL, SDL_FLIP_NONE);
-	}
-}
+#include "gameplay.h"
 
 void end_turn(void);
 void move_human(EntId eid, TileCoords dst_pos);
 
 void cycle_ent_sel_through_ents_in_tile(void);
 
-/* The human (given by its entity ID `eid`) will build a building (of the given type)
- * on the tile at the given coords. */
-void have_human_to_build(EntId eid, BuildingType building_type, TileCoords tc) {
-	new_building(building_type, tc);
-	Ent* ent = get_ent(eid);
-	assert(ent != NULL && ent->type == ENT_HUMAN);
-	ent->human.already_moved_this_turn = true;
-	refresh_selected_tile_ui();
-	DA_EMPTY_LEAK(&g_action_da_on_tcs);
-	action_menu_refresh();
-}
 
-/* The human (given by its entity ID `eid`) will perform the given action
- * on the tile at the given coords. */
-void have_human_to_act(EntId eid, Action const* action, TileCoords tc) {
-	Ent* ent = get_ent(eid);
-	assert(ent != NULL && ent->type == ENT_HUMAN);
-	assert(action != NULL);
-	switch (action->type) {
-		case ACTION_MOVE:
-			move_human(g_sel_ent_id, tc);
-		break;
-		case ACTION_BUILD: {
-			have_human_to_build(g_sel_ent_id, action->build.building_type, tc);
-		break; }
-		default: assert(false);
-	}
-}
 
-void click(WinCoords wc) {
-	/* The user clicked on the pixel at `wc`.
-	 * First we forward this click to the UI in case it lands on a widget. */
-	bool ui_used_the_click = ui_click(wc);
-	if (ui_used_the_click) {
-		return;
-	}
 
-	/* The UI said the click was not for them, so it lands on the map. */
-	TileCoords tc = window_pixel_to_tile_coords(wc);
-	if (tile_is_available(tc)) {
-		/* A movable human is selected and the click landed on a tile to which
-		 * the human can move or act. */
-		assert(g_sel_ent_exists);
-		Action const* action = action_menu_selection();
-		have_human_to_act(g_sel_ent_id, action, tc);
-	} else if (g_sel_tile_exists && tile_coords_eq(g_sel_tile_coords, tc)) {
-		/* The click landed on the selected tile. */
-		Tile const* tile = get_tile(g_sel_tile_coords);
-		if (get_tile_real_ent_count(tile) > 0) {
-			cycle_ent_sel_through_ents_in_tile();
-		} else {
-			ui_unselect_tile();
-		}
-	} else if (tile_coords_are_valid(tc)) {
-		ui_unselect_ent();
-		ui_select_tile(tc);
-	} else {
-		ui_unselect_tile();
-	}
-}
 
-bool ent_is_playing(EntId eid) {
-	Ent* ent = get_ent(eid);
-	return
-		ent != NULL &&
-		ent->type == ENT_HUMAN &&
-		ent->human.faction == g_faction_currently_playing;
-}
-
-bool ent_can_move(EntId eid) {
-	Ent* ent = get_ent(eid);
-	return
-		ent_is_playing(eid) &&
-		(!ent->human.already_moved_this_turn);
-}
-
-bool keycode_is_arrow(SDL_Keycode keycode) {
-	return
-		keycode == SDLK_DOWN ||
-		keycode == SDLK_UP ||
-		keycode == SDLK_RIGHT ||
-		keycode == SDLK_LEFT;
-}
-
-typedef struct DxDy {int dx, dy;} DxDy;
-
-TileCoords tc_add_dxdy(TileCoords tc, DxDy dxdy) {
-	return (TileCoords){tc.x + dxdy.dx, tc.y + dxdy.dy};
-}
-
-DxDy arrow_keycode_to_dxdy(SDL_Keycode keycode) {
-	switch (keycode) {
-		case SDLK_DOWN:  return (DxDy){0, +1};
-		case SDLK_UP:    return (DxDy){0, -1};
-		case SDLK_RIGHT: return (DxDy){+1, 0};
-		case SDLK_LEFT:  return (DxDy){-1, 0};
-		default: assert(false); exit(EXIT_FAILURE);
-	}
-}
-
-DxDy cardinal_to_dxdy(CardinalType card) {
-	switch (card) {
-		case SOUTH: return (DxDy){0, +1};
-		case NORTH: return (DxDy){0, -1};
-		case WEST:  return (DxDy){+1, 0};
-		case EAST:  return (DxDy){-1, 0};
-		default: assert(false); exit(EXIT_FAILURE);
-	}
-}
-
-void cycle_ent_sel_through_ents_in_tile(void) {
-	if (!g_sel_tile_exists) return;
-	Tile* sel_tile = get_tile(g_sel_tile_coords);
-
-	/* Here we only want to select entities that belong to the currently playing faction.
-	 * `first` is the first of such entities (if any).
-	 * `first_after_sel` is the first of such that follows the currently selected entity (if any). */
-	EntId first = EID_NULL;
-	bool sel_ent_is_in_sel_tile = false;
-	EntId first_after_sel = EID_NULL;
-	for (int i = 0; i < sel_tile->ents.len; i++) {
-		EntId eid = sel_tile->ents.arr[i];
-		if (!ent_is_playing(eid)) continue;
-		if (eid_null(first)) {
-			first = eid;
-		}
-		if (eid_eq(g_sel_ent_id, eid)) {
-			sel_ent_is_in_sel_tile = true;
-		} else if (sel_ent_is_in_sel_tile &&
-			eid_null(first_after_sel)
-		) {
-			first_after_sel = eid;
-		}
-	}
-
-	/* Now we just select the next of such entities if that makes sense,
-	 * else we just select the first (if any). */
-	if (sel_ent_is_in_sel_tile && eid_null(first_after_sel)) {
-		ui_select_ent(first);
-	} else if (sel_ent_is_in_sel_tile) {
-		ui_select_ent(first_after_sel);
-	} else if (!eid_null(first)) {
-		ui_select_ent(first);
-	} else {
-		ui_unselect_ent();
-	}
-}
-
-SDL_Rect tile_rect(TileCoords tc) {
-	float tile_render_size = TILE_SIZE * g_camera.zoom;
-	return (SDL_Rect){
-		.x = tc.x * tile_render_size - g_camera.pos.x,
-		.y = tc.y * tile_render_size - g_camera.pos.y,
-		.w = ceilf(tile_render_size),
-		.h = ceilf(tile_render_size)};
-}
 
 int get_tile_real_ent_count(Tile const* tile) {
 	int true_ent_count = 0;
@@ -261,179 +36,6 @@ int get_tile_real_ent_count(Tile const* tile) {
 	return true_ent_count;
 }
 
-/* Is grid line display enabled? */
-bool g_render_lines = false;
-
-Uint64 g_time_ms = 0;
-
-
-void render_map(void) {
-	float tile_render_size = TILE_SIZE * g_camera.zoom;
-
-	/* Draw tile terrain. */
-	for (int i = 0; i < N_TILES; ++i) {
-		TileCoords tc = {.x = i % g_map_w, .y = i / g_map_w};
-		Tile const* tile = get_tile(tc);
-		SDL_Rect rect = tile_rect(tc);
-		render_tile_ground(tile->type, rect);
-	}
-
-	if (g_render_lines) {
-		/* Draw grid lines. */
-		SDL_SetRenderDrawColor(g_renderer, 0, 0, 0, 255);
-		for (int y = 0; y < g_map_w+1; ++y) {
-			WinCoords a = tile_coords_to_window_pixel((TileCoords){0, y});
-			WinCoords b = tile_coords_to_window_pixel((TileCoords){g_map_w, y});
-			for (int i = 0; i < 2; i++) {
-				SDL_RenderDrawLine(g_renderer, a.x, a.y - i, b.x, b.y - i);
-			}
-		}
-		for (int x = 0; x < g_map_h+1; ++x) {
-			WinCoords a = tile_coords_to_window_pixel((TileCoords){x, 0});
-			WinCoords b = tile_coords_to_window_pixel((TileCoords){x, g_map_h});
-			for (int i = 0; i < 2; i++) {
-				SDL_RenderDrawLine(g_renderer, a.x - i, a.y, b.x - i, b.y);
-			}
-		}
-	}
-
-	/* Draw selection rect around the selected tile (if any).
-	 * We do this after all the tile terrains so that terrain drawn after the
-	 * selection rectangle won't cover a part of it. */
-	if (g_sel_tile_exists) {
-		SDL_Rect rect = tile_rect(g_sel_tile_coords);
-		SDL_SetRenderDrawColor(g_renderer, 255, 255, 0, 255);
-		for (int i = 0; i < 3; i++) {
-			SDL_RenderDrawRect(g_renderer, &rect);
-			rect.x += 1; rect.y += 1; rect.w -= 2; rect.h -= 2;
-		}
-	}
-
-	for (int i = 0; i < g_action_da_on_tcs.len; i++) {
-		SDL_Rect rect = tile_rect(g_action_da_on_tcs.arr[i].tc);
-		SDL_Point mouse;
-		SDL_GetMouseState(&mouse.x, &mouse.y);
-		if (SDL_PointInRect(&mouse, &rect)) {
-			SDL_SetRenderDrawColor(g_renderer, 0, 255, 255, 255);
-		} else {
-			SDL_SetRenderDrawColor(g_renderer, 0, 0, 255, 255);
-		}
-		for (int i = 0; i < 2; i++) {
-			rect.x += 1; rect.y += 1; rect.w -= 2; rect.h -= 2;
-			SDL_RenderDrawRect(g_renderer, &rect);
-		}
-	}
-
-	/* Draw entities, buildings and flows.
-	 * We do this after the tile backgrounds so that these can cover a part
-	 * of neighboring tiles. */
-	for (int i = 0; i < N_TILES; ++i) {
-		TileCoords tc = {.x = i % g_map_w, .y = i / g_map_w};
-		Tile const* tile = get_tile(tc);
-		SDL_Rect rect = tile_rect(tc);
-
-		/* Draw flows. */
-		for (int cable_i = 0; cable_i < tile->cable_count; cable_i++){
-			Cable* cable = tile->cables[cable_i];
-			assert(cable != NULL);
-			render_tile_cable(cable, rect);
-		}
-
-		/* Draw building. */
-		if (tile->building != NULL){
-			render_tile_building(tile->building, rect);
-		}
-
-		/* Draw entities.
-		 * Note that `tile->ents.arr` can contain empty cells, so we cannot count entities
-		 * with this dynamic array and must use `real_ent_count` and `real_ent_i` instead. */
-		int real_ent_count = get_tile_real_ent_count(tile);
-		int real_ent_i = -1;
-		for (int ent_i = 0; ent_i < tile->ents.len; ent_i++) {
-			EntId eid = (tile->ents.arr[ent_i]);
-			Ent* ent = get_ent(eid);
-			if (ent == NULL) continue;
-			real_ent_i++;
-
-			/* Entity position on screen. */
-			int ex = (float)(real_ent_i+1) / (float)(real_ent_count+1)
-				* tile_render_size;
-			int ey = (1.0f - (float)(real_ent_i+1) / (float)(real_ent_count+1))
-				* tile_render_size;
-
-			/* Account for an ongoing animation (if any). */
-			if (ent->anim != NULL &&
-				ent->anim->time_beginning <= g_time_ms && g_time_ms < ent->anim->time_end
-			) {
-				float interpolation =
-					(float)(g_time_ms - ent->anim->time_beginning) /
-					(float)(ent->anim->time_end - ent->anim->time_beginning);
-				/* The offset is maximal at the beginning of the animation (when `interpolation`
-				 * is zero) and tends toward zero as `interpolation` tends toward one. */
-				ex -= ent->anim->offset_beginning_x * tile_render_size * (1.0f - interpolation);
-				ey -= ent->anim->offset_beginning_y * tile_render_size * (1.0f - interpolation);
-			}
-
-			switch (ent->type) {
-				case ENT_HUMAN: {
-					int ew = 0.3f * tile_render_size;
-					int eh = 0.4f * tile_render_size;
-
-					/* Draw human sprite. */
-					SDL_Rect ent_rect = {
-						rect.x + ex - ew / 2.0f, rect.y + ey - eh / 2.0f, ew, eh};
-					if (g_sel_ent_exists && eid_eq(g_sel_ent_id, eid)) {
-						/* If the entity is selected, then make it a bit bigger
-						 * its fixed point is its bottom middle point as its looks better that way. */
-						ent_rect.x -= 2; ent_rect.w += 4;
-						ent_rect.y -= 4; ent_rect.h += 4;
-					}
-					render_human(ent_rect);
-
-					/* Draw faction color square. */
-					SDL_Color color = g_faction_spec_table[ent->human.faction].color;
-					SDL_SetRenderDrawColor(g_renderer, color.r, color.g, color.b, 255);
-					#define FACTION_SIDE 8
-					SDL_Rect faction_rect = {
-						ent_rect.x + ent_rect.w/2 - FACTION_SIDE/2,
-						ent_rect.y - FACTION_SIDE/2 - FACTION_SIDE,
-						FACTION_SIDE, FACTION_SIDE};
-					#undef FACTION_SIDE
-					if (g_sel_ent_exists && eid_eq(g_sel_ent_id, eid)) {
-						faction_rect.y -= 2;
-					}
-					SDL_RenderFillRect(g_renderer, &faction_rect);
-
-					/* Draw some black and white boarders around the faction color square of the
-					 * selected entity (if any). */
-					if (g_sel_ent_exists && eid_eq(g_sel_ent_id, eid)) {
-						SDL_SetRenderDrawColor(g_renderer, 255, 255, 255, 255);
-						for (int i = 0; i < 2; i++) {
-							faction_rect.x-=1; faction_rect.y-=1; faction_rect.w+=2; faction_rect.h+=2;
-							SDL_RenderDrawRect(g_renderer, &faction_rect);
-						}
-						SDL_SetRenderDrawColor(g_renderer, 0, 0, 0, 255);
-						for (int i = 0; i < 2; i++) {
-							faction_rect.x-=1; faction_rect.y-=1; faction_rect.w+=2; faction_rect.h+=2;
-							SDL_RenderDrawRect(g_renderer, &faction_rect);
-						}
-					}
-				break; }
-
-				case ENT_TEST_BLOCK: {
-					int ew = 0.2f * tile_render_size;
-					int eh = 0.2f * tile_render_size;
-					SDL_Rect ent_rect = {rect.x + ex - ew / 2.0f, rect.y + ey - eh / 2.0f, ew, eh};
-					SDL_Color color = ent->test_block.color;
-					SDL_SetRenderDrawColor(g_renderer, color.r, color.g, color.b, 255);
-					SDL_RenderFillRect(g_renderer, &ent_rect);
-				break; }
-
-				default: assert(false);
-			}
-		}
-	}
-}
 
 int main(int argc, char const** argv) {
 	/* Parse command line arguments. */
